ex 4 9:30
