+++ src/components/escena/escena.js
@@ -1,69 +1,37 @@
 //importar react
 import React from "react";
 import { EscenaStyled } from "./escena.styled";
-import { useState } from "react";
 
-
-
-
-
-    
-    
-
-    
 
 
 
 export default function Escena(props) {
 
-
-    
-
-
-
-
-
     return (
+
+
 
         <div>
 
-           
-        
-
-
-
             
 
+                <EscenaStyled 
 
-            
-
-            
-
-
-                <EscenaStyled
-            
-                    fons={false}
-            
+                fons={true}
+                
+                
                 >
+
+                    {props.vermell}
 
                     {props.frase}
 
-                </EscenaStyled>
-
-
-            
+                </EscenaStyled> 
 
 
         </div>
 
-        
-
-        
-
-        
-
     )
-
 
 }
